# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExamenTAP API",
        "description": "Documentación de la API REST del backend ExamenTAP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/perfiles": {
            "get": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Lista perfiles con paginación, búsqueda y ordenamiento",
                "description": "Devuelve un listado paginado de perfiles. Permite buscar por código o nombre. Requiere el permiso \"consultar\" en la sección \"perfiles\".",
                "operationId": "perfilesIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "id",
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerfilListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Crea un nuevo perfil",
                "description": "Crea un perfil validando que el nombre sea único y define la matriz de permisos por sección. Requiere el permiso \"crear\" en la sección \"perfiles\".",
                "operationId": "perfilesStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePerfilRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perfil creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerfilResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para crear perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfiles/{perfil}": {
            "get": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Obtiene el detalle de un perfil",
                "description": "Devuelve la información completa de un perfil, incluyendo la matriz de secciones y permisos asociada. Requiere el permiso \"consultar\" en la sección \"perfiles\".",
                "operationId": "perfilesShow",
                "parameters": [
                    {
                        "name": "perfil",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de perfil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerfilResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Actualiza un perfil existente",
                "description": "Actualiza el nombre y la matriz de permisos por sección de un perfil. Requiere el permiso \"editar\" en la sección \"perfiles\".",
                "operationId": "perfilesUpdate",
                "parameters": [
                    {
                        "name": "perfil",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePerfilRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PerfilResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para editar perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Elimina un perfil",
                "description": "Elimina físicamente el perfil de Firebase Firestore. Requiere el permiso \"eliminar\" en la sección \"perfiles\".",
                "operationId": "perfilesDestroy",
                "parameters": [
                    {
                        "name": "perfil",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permiso para eliminar perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfiles/export/pdf": {
            "get": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Exporta el listado de perfiles a PDF",
                "description": "Genera un PDF con Código, Nombre y Fecha de creación. Requiere el permiso \"consultar\" en la sección \"perfiles\".",
                "operationId": "perfilesExportPdf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo PDF",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfiles/export/excel": {
            "get": {
                "tags": [
                    "Perfiles"
                ],
                "summary": "Exporta el listado de perfiles a Excel",
                "description": "Genera un archivo XLSX con Código, Nombre y Fecha de creación. Requiere el permiso \"consultar\" en la sección \"perfiles\".",
                "operationId": "perfilesExportExcel",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo Excel",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    },
                    "403": {
                        "description": "Sin permiso para consultar perfiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Las credenciales proporcionadas son incorrectas."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El correo electrónico es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "PerfilListResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Perfiles obtenidos correctamente."
                    },
                    "data": {
                        "properties": {
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Perfil"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PerfilResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Perfil creado correctamente."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Perfil"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Pagination": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
            "StorePerfilRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "secciones": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "seccion": {
                                    "type": "string",
                                    "example": "usuarios"
                                },
                                "crear": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "consultar": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "editar": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "eliminar": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UpdatePerfilRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "secciones": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "seccion": {
                                    "type": "string",
                                    "example": "usuarios"
                                },
                                "crear": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "consultar": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "editar": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "eliminar": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}